% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.case5.R
\name{wrapper.case5}
\alias{wrapper.case5}
\title{Returns the future value of annuity payments (fv) with inflation adjusted.}
\usage{
wrapper.case5(rate = 0, inflation = 0, nper = 1, pv = 0)
}
\arguments{
\item{rate}{The interest rate per period. Default is zero. Must be entered as decimal}

\item{inflation}{The inflation rate per period. Default is zero. Must be entered as decimal}

\item{nper}{The total number of payment periods. Default is one period}

\item{pv}{The present value of single payment made today. Default is assumed to be zero. Must be entered as a negative number}
}
\description{
Returns the future value of annuity payments (fv) with inflation adjusted.
}
\examples{
wrapper.case5(rate=0.04,inflation=0.02,nper=35,pv=-1000)
}
\seealso{
\code{\link{fv.single}}

\code{\link{fv.annuity}}
}
