% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.case8.R
\name{wrapper.case8}
\alias{wrapper.case8}
\title{Returns the future value (fv) with withdrawals.}
\usage{
wrapper.case8(rate = 0, inflation = 0, nper_savings = 1,
  nper_withdrawals = 0, pmt = 0)
}
\arguments{
\item{rate}{The interest rate per period. Default is zero. Must be entered as decimal}

\item{inflation}{The inflation rate per period. Default is zero. Must be entered as decimal}

\item{nper_savings}{The total number of payment periods. Default is one period}

\item{nper_withdrawals}{The total number of withdraws periods. Default is zero period}

\item{pmt}{The payment made each period (annuity). Must be entered as a negative number.}
}
\description{
Returns the future value (fv) with withdrawals.
}
\examples{
wrapper.case8(rate=0.04,inflation=0.02,nper_savings=35,nper_withdrawals=30,pmt=-1000)
}
\seealso{
\code{\link{fv}}

\code{\link{pmt}}
}
