% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.case9.R
\name{wrapper.case9}
\alias{wrapper.case9}
\title{Returns the PV(payments), adj. for inflation.}
\usage{
wrapper.case9(rate = 0, inflation = 0, nper = 1, pmt = 0)
}
\arguments{
\item{rate}{The interest rate per period. Default is zero. Must be entered as decimal}

\item{inflation}{The inflation rate per period. Default is zero. Must be entered as decimal}

\item{nper}{The total number of payment periods. Default is one period}

\item{pmt}{The payment made each period (annuity). Must be entered as a negative number.}
}
\description{
Returns the PV(payments), adj. for inflation.
}
\examples{
wrapper.case9(rate=0.04,inflation=0.02,nper=35,pmt=-1000)
}
\seealso{
\code{\link{rate.real}}

\code{\link{pv}}
}
